+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+- support for cc, pitch-bend and bank-select
+
 ## [0.3.0] - 2021-12-24
 
 ### Changed
@@ -22,6 +28,7 @@
 
 Initial release.
 
+[Unreleased]: https://github.com/omkamra/sequencer-fluidsynth/compare/0.3.0...HEAD
 [0.3.0]: https://github.com/omkamra/sequencer-fluidsynth/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/omkamra/sequencer-fluidsynth/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/omkamra/sequencer-fluidsynth/compare/0.1.0...0.2.0